Added Signup/Signin cookies, app.js cleanup
