+++ package.json
@@ -10,10 +10,12 @@
   "license": "ISC",
   "dependencies": {
     "bcryptjs": "^2.4.3",
+    "cookie-parser": "^1.4.6",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "googleapis": "^118.0.0",
     "hbs": "^4.2.0",
+    "jsonwebtoken": "^9.0.2",
     "mongoose": "^7.0.4",
     "nodemailer": "^6.9.2",
     "nodemon": "^2.0.20",
